ml-service/requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
google-generativeai
python-multipart